+++ src/main/java/com/github/theredbrain/rpginventory/mixin/screen/PlayerScreenHandlerOffHandSlotMixin.java
@@ -33,7 +33,7 @@ public abstract class PlayerScreenHandlerOffHandSlotMixin extends Slot {
 
 	@Override
 	public boolean isEnabled() {
-		return !((DuckPlayerEntityMixin) this.field_39410).rpginventory$isOffhandStackSheathed();
+		return !RPGInventory.isHandSlotOverhaulActive() || !((DuckPlayerEntityMixin) this.field_39410).rpginventory$isOffhandStackSheathed();
 	}
 
 	@Override
@@ -50,7 +50,9 @@ public abstract class PlayerScreenHandlerOffHandSlotMixin extends Slot {
 		Optional<RegistryEntry.Reference<StatusEffect>> wilderness_status_effect = Registries.STATUS_EFFECT.getEntry(serverConfig.statusEffects.wilderness_status_effect_identifier.get());
 		boolean hasWildernessEffect = wilderness_status_effect.isPresent() && this.field_39410.hasStatusEffect(wilderness_status_effect.get());
 
-		return (EquipmentSlot.OFFHAND == this.field_39410.getPreferredEquipmentSlot(stack) || stack.isIn(Tags.OFFHAND_ITEMS) || !serverConfig.handSlotOverhaul.are_hand_items_restricted_to_item_tags.get() || !RPGInventory.isHandSlotOverhaulActive()) && ItemUtils.isOwnedByPlayer(stack, this.field_39410.getGameProfile()) && (hasCivilisationEffect || this.field_39410.isCreative() || (bl && !hasWildernessEffect)) && !((DuckPlayerEntityMixin) this.field_39410).rpginventory$isOffhandStackSheathed();
+		boolean handSlotOverhaulIsInactive = !RPGInventory.isHandSlotOverhaulActive();
+
+		return (EquipmentSlot.OFFHAND == this.field_39410.getPreferredEquipmentSlot(stack) || stack.isIn(Tags.OFFHAND_ITEMS) || !serverConfig.handSlotOverhaul.are_hand_items_restricted_to_item_tags.get() || handSlotOverhaulIsInactive) && (handSlotOverhaulIsInactive || !((DuckPlayerEntityMixin) this.field_39410).rpginventory$isOffhandStackSheathed()) && ItemUtils.isOwnedByPlayer(stack, this.field_39410.getGameProfile()) && (hasCivilisationEffect || this.field_39410.isCreative() || (bl && !hasWildernessEffect));
 	}
 
 	@Override